pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pjm-south-lmp-forecast"
version = "0.1.0"
description = "Hourly Day-Ahead LMP forecasting for PJM SOUTH aggregate node"
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "lightgbm>=4.0.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "optuna>=3.3.0",
    "shap>=0.43.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyarrow>=12.0.0",
    "tabulate>=0.9.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov"]

[project.scripts]
pjm-forecast = "cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
